Update event script
